# src/main/java/com/walmart/ticket/model/FindAndHold.java
package com.walmart.ticket.model;

/**
 * Created by bmaria001c on 5/3/16.
 */
public class FindAndHold {
    private String emailId;
    private int minLevel;
    private int maxLevel;
    private int numberOfSeats;
    public int getMaxLevel() {
        return maxLevel;
    }

    public void setMaxLevel(int maxLevel) {
        this.maxLevel = maxLevel;
    }

    public int getNumberOfSeats() {
        return numberOfSeats;
    }

    public void setNumberOfSeats(int numberOfSeats) {
        this.numberOfSeats = numberOfSeats;
    }

    public int getMinLevel() {
        return minLevel;
    }

    public void setMinLevel(int minLevel) {
        this.minLevel = minLevel;
    }

    public String getEmailId() {
        return emailId;
    }

    public void setEmailId(String emailId) {
        this.emailId = emailId;
    }


}